+++ src/frontend/PayerProjects/PayerProjectSetup/PayerProjectSetup.js
@@ -1,9 +1,6 @@
 import React from 'react'
-import { useQuery } from '@apollo/react-hooks'
 import styled from '@emotion/styled'
 
-import { GET_SINGLE_PAYER_PROJECT } from 'frontend/api/queries'
-import Spinner from 'frontend/components/Spinner'
 import ProjectInfo from './ProjectInfo'
 import ProjectContentConfiguration from './ProjectContentConfiguration'
 
@@ -15,21 +12,9 @@ const Wrapper = styled.div({
 })
 
 const PayerProjectSetup = props => {
-  const { projectId } = props.match.params
-
-  const { data, loading } = useQuery(
-    GET_SINGLE_PAYER_PROJECT,
-    {
-      variables: { projectId },
-    })
-
-  if (loading) return <Spinner />
-
-  const { singlePayerProject: project } = data
-
   return (
     <Wrapper>
-      <ProjectInfo name={project.name} />
+      <ProjectInfo name={props.projectName} />
       <ProjectContentConfiguration />
     </Wrapper>
   )